+++ packages/base/src/decorators.ts
@@ -3,6 +3,8 @@ import event from "./decorators/event.js";
 import eventStrict from "./decorators/event-strict.js";
 import property from "./decorators/property.js";
 import slot from "./decorators/slot.js";
+import bound from "./decorators/bound.js";
+import i18n from "./decorators/i18n.js";
 
 export {
 	customElement,
@@ -10,4 +12,6 @@ export {
 	eventStrict,
 	property,
 	slot,
+	bound,
+	i18n,
 };