+++ packages/main/src/CalendarLegend.ts
@@ -1,8 +1,9 @@
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
+import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import ItemNavigation from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
 import NavigationMode from "@ui5/webcomponents-base/dist/types/NavigationMode.js";
@@ -10,10 +11,12 @@ import {
 	isDown,
 	isUp,
 } from "@ui5/webcomponents-base/dist/Keys.js";
+import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import ItemNavigationBehavior from "@ui5/webcomponents-base/dist/types/ItemNavigationBehavior.js";
 import CalendarLegendItemType from "./types/CalendarLegendItemType.js";
-import CalendarLegendTemplate from "./generated/templates/CalendarLegendTemplate.lit.js";
+import CalendarLegendTemplate from "./CalendarLegendTemplate.js";
 import CalendarLegendItem from "./CalendarLegendItem.js";
+import { CAL_LEGEND_ROLE_DESCRIPTION } from "./generated/i18n/i18n-defaults.js";
 
 // Styles
 import CalendarLegendCss from "./generated/themes/CalendarLegend.css.js";
@@ -41,22 +44,23 @@ type CalendarLegendItemSelectionChangeEventDetail = {
  */
 @customElement({
 	tag: "ui5-calendar-legend",
-	renderer: litRender,
+	renderer: jsxRenderer,
 	styles: CalendarLegendCss,
 	template: CalendarLegendTemplate,
 	dependencies: [CalendarLegendItem],
 })
-@event("_calendar-legend-selection-change", {
+@event("calendar-legend-selection-change", {
 	bubbles: true,
 })
-@event("_calendar-legend-focus-out", {
+@event("calendar-legend-focus-out", {
 	bubbles: true,
 })
 class CalendarLegend extends UI5Element {
 	eventDetails!: {
-		"_calendar-legend-selection-change": CalendarLegendItemSelectionChangeEventDetail,
-		"_calendar-legend-focus-out": void,
-	}
+		"calendar-legend-selection-change": CalendarLegendItemSelectionChangeEventDetail,
+		"calendar-legend-focus-out": void,
+	};
+
 	/**
 	 * Hides the Today item in the legend.
 	 * @default false
@@ -104,6 +108,9 @@ class CalendarLegend extends UI5Element {
 	_itemNavigation!: ItemNavigation;
 	_lastFocusedItemIndex: number | null;
 
+	@i18n("@ui5/webcomponents")
+	static i18nBundle: I18nBundle;
+
 	constructor() {
 		super();
 
@@ -135,14 +142,14 @@ class CalendarLegend extends UI5Element {
 	_onFocusIn(e: FocusEvent) {
 		const target = e.target as CalendarLegendItem;
 
-		this.fireDecoratorEvent("_calendar-legend-selection-change", {
+		this.fireDecoratorEvent("calendar-legend-selection-change", {
 			item: target,
 		});
 		this._lastFocusedItemIndex = this.focusableElements.indexOf(target);
 	}
 
 	_onFocusOut() {
-		this.fireDecoratorEvent("_calendar-legend-focus-out");
+		this.fireDecoratorEvent("calendar-legend-focus-out");
 	}
 
 	_onItemKeyDown(e: KeyboardEvent) {
@@ -186,13 +193,17 @@ class CalendarLegend extends UI5Element {
 
 	get defaultItemsMapping() {
 		const typeMapping = [
-			{ type: [CalendarLegendItemType.Today], hide: this.hideToday },
-			{ type: [CalendarLegendItemType.Selected], hide: this.hideSelectedDay },
-			{ type: [CalendarLegendItemType.Working], hide: this.hideWorkingDay },
-			{ type: [CalendarLegendItemType.NonWorking], hide: this.hideNonWorkingDay },
+			{ type: CalendarLegendItemType.Today, hide: this.hideToday },
+			{ type: CalendarLegendItemType.Selected, hide: this.hideSelectedDay },
+			{ type: CalendarLegendItemType.Working, hide: this.hideWorkingDay },
+			{ type: CalendarLegendItemType.NonWorking, hide: this.hideNonWorkingDay },
 		];
 
 		return typeMapping;
+	}
+
+	get _roleDescription() {
+		return CalendarLegend.i18nBundle.getText(CAL_LEGEND_ROLE_DESCRIPTION);
 	}
 }