+++ packages/main/src/ListItem.ts
@@ -3,14 +3,14 @@ import {
 	isSpace, isEnter, isDelete, isF2,
 } from "@ui5/webcomponents-base/dist/Keys.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import getActiveElement from "@ui5/webcomponents-base/dist/util/getActiveElement.js";
 import { getFirstFocusableElement } from "@ui5/webcomponents-base/dist/util/FocusableElements.js";
-import type { AccessibilityAttributes, PassiveEventListenerObject } from "@ui5/webcomponents-base/dist/types.js";
+import type { AccessibilityAttributes, AriaRole, AriaHasPopup } from "@ui5/webcomponents-base";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
-import type AriaHasPopup from "@ui5/webcomponents-base/dist/types/AriaHasPopup.js";
 import "@ui5/webcomponents-icons/dist/decline.js";
 import "@ui5/webcomponents-icons/dist/edit.js";
 import Highlight from "./types/Highlight.js";
@@ -50,13 +50,13 @@ type SelectionRequestEventDetail = {
 }
 
 type AccInfo = {
-	role?: string;
+	role?: AriaRole | undefined;
 	ariaExpanded?: boolean;
 	ariaLevel?: number;
 	ariaLabel: string;
 	ariaLabelRadioButton: string;
 	ariaSelectedText?: string;
-	ariaHaspopup?: `${Lowercase<AriaHasPopup>}`;
+	ariaHaspopup?: `${AriaHasPopup}`;
 	posinset?: number;
 	setsize?: number;
 	ariaSelected?: boolean;
@@ -80,6 +80,7 @@ type ListItemAccessibilityAttributes = Pick<AccessibilityAttributes, "hasPopup" 
  */
 @customElement({
 	languageAware: true,
+	renderer: jsxRenderer,
 	styles: [
 		ListItemBase.styles,
 		listItemAdditionalTextCss,
@@ -98,13 +99,13 @@ type ListItemAccessibilityAttributes = Pick<AccessibilityAttributes, "hasPopup" 
 @event("detail-click", {
 	bubbles: true,
 })
-@event("_selection-requested", {
+@event("selection-requested", {
 	bubbles: true,
 })
 abstract class ListItem extends ListItemBase {
 	eventDetails!: ListItemBase["eventDetails"] & {
 		"detail-click": { item: ListItem, selected: boolean };
-		"_selection-requested": SelectionRequestEventDetail,
+		"selection-requested": SelectionRequestEventDetail,
 	}
 	/**
 	 * Defines the visual indication and behavior of the list items.
@@ -190,6 +191,9 @@ abstract class ListItem extends ListItemBase {
 	accessibleRole: `${ListItemAccessibleRole}` = "ListItem";
 
 	@property()
+	_forcedAccessibleRole?: string;
+
+	@property()
 	_selectionMode: `${ListSelectionMode}` = "None";
 
 	/**
@@ -205,7 +209,6 @@ abstract class ListItem extends ListItemBase {
 
 	deactivateByKey: (e: KeyboardEvent) => void;
 	deactivate: () => void;
-	_ontouchstart: PassiveEventListenerObject;
 	// used in template, implemented in TreeItemBase, ListItemStandard
 	accessibleName?: string;
 	// used in ListItem template but implemented in TreeItemBase
@@ -229,15 +232,6 @@ abstract class ListItem extends ListItemBase {
 			if (this.active) {
 				this.active = false;
 			}
-		};
-
-		const handleTouchStartEvent = () => {
-			this._onmousedown();
-		};
-
-		this._ontouchstart = {
-			handleEvent: handleTouchStartEvent,
-			passive: true,
 		};
 	}
 
@@ -358,20 +352,20 @@ abstract class ListItem extends ListItemBase {
 	 * Called when selection components in Single (ui5-radio-button)
 	 * and Multi (ui5-checkbox) selection modes are used.
 	 */
-	onMultiSelectionComponentPress(e: MouseEvent) {
+	onMultiSelectionComponentPress(e: CustomEvent) {
 		if (this.isInactive) {
 			return;
 		}
 
-		this.fireDecoratorEvent("_selection-requested", { item: this, selected: (e.target as CheckBox).checked, selectionComponentPressed: true });
+		this.fireDecoratorEvent("selection-requested", { item: this, selected: (e.target as CheckBox).checked, selectionComponentPressed: true });
 	}
 
-	onSingleSelectionComponentPress(e: MouseEvent) {
+	onSingleSelectionComponentPress(e: CustomEvent) {
 		if (this.isInactive) {
 			return;
 		}
 
-		this.fireDecoratorEvent("_selection-requested", { item: this, selected: !(e.target as RadioButton).checked, selectionComponentPressed: true });
+		this.fireDecoratorEvent("selection-requested", { item: this, selected: !(e.target as RadioButton).checked, selectionComponentPressed: true });
 	}
 
 	activate() {
@@ -381,7 +375,7 @@ abstract class ListItem extends ListItemBase {
 	}
 
 	onDelete() {
-		this.fireDecoratorEvent("_selection-requested", { item: this, selectionComponentPressed: false });
+		this.fireDecoratorEvent("selection-requested", { item: this, selectionComponentPressed: false });
 	}
 
 	onDetailClick() {
@@ -393,6 +387,9 @@ abstract class ListItem extends ListItemBase {
 			return;
 		}
 		super.fireItemPress(e);
+		if (document.activeElement !== this) {
+			this.focus();
+		}
 	}
 
 	get isInactive() {
@@ -457,7 +454,7 @@ abstract class ListItem extends ListItemBase {
 	}
 
 	get listItemAccessibleRole() {
-		return this.accessibleRole.toLowerCase();
+		return (this._forcedAccessibleRole || this.accessibleRole.toLowerCase()) as AriaRole | undefined;
 	}
 
 	get ariaSelectedText() {