+++ packages/main/src/ResponsivePopover.ts
@@ -6,17 +6,14 @@ import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
 
 import { RESPONSIVE_POPOVER_CLOSE_DIALOG_BUTTON } from "./generated/i18n/i18n-defaults.js";
 
-import ResponsivePopoverTemplate from "./generated/templates/ResponsivePopoverTemplate.lit.js";
-import type { PopupBeforeCloseEventDetail } from "./Popup.js";
+import ResponsivePopoverTemplate from "./ResponsivePopoverTemplate.js";
 import Popover from "./Popover.js";
 import Dialog from "./Dialog.js";
 import Button from "./Button.js";
-import "@ui5/webcomponents-icons/dist/decline.js";
 
 // Styles
 import ResponsivePopoverCss from "./generated/themes/ResponsivePopover.css.js";
-
-type ResponsivePopoverBeforeCloseEventDetail = PopupBeforeCloseEventDetail;
+import type { PopupBeforeCloseEventDetail } from "./Popup.js";
 
 /**
  * @class
@@ -96,6 +93,10 @@ class ResponsivePopover extends Popover {
 		if (!isPhone()) {
 			return super._show();
 		}
+	}
+
+	_dialogCloseButtonClick() {
+		this.closePopup();
 	}
 
 	/**
@@ -187,6 +188,3 @@ class ResponsivePopover extends Popover {
 ResponsivePopover.define();
 
 export default ResponsivePopover;
-export type {
-	ResponsivePopoverBeforeCloseEventDetail,
-};